# app/internal/api/docs/swagger.yaml
basePath: /
definitions:
...
  model.CreateGoodRequest:
    properties:
      category_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      color_code:
        example: '#FF5733'
        type: string
      cook_time:
        example: 30
        type: integer
      department_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      description:
        example: Classic Italian pizza with tomatoes and mozzarella
        type: string
      description_i18n:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      name:
        example: Pizza Margherita
        type: string
      name_i18n:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      picture_url:
        example: https://example.com/pizza-margherita.jpg
        type: string
      price:
        example: "15000.00"
        type: string
    required:
    - name
    - price
    type: object
...
  model.ErrorResponse:
    properties:
      message:
        example: error message
        type: string
    type: object
...
  model.GoodResponse:
    properties:
      category_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      color_code:
        example: '#FF5733'
        type: string
      cook_time:
        example: 30
        type: integer
      created_at:
        example: "2022-01-01T00:00:00Z"
        type: string
      department_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      description:
        example: Classic Italian pizza with tomatoes and mozzarella
        type: string
      description_i18n:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      name:
        example: Pizza Margherita
        type: string
      name_i18n:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      picture_url:
        example: https://example.com/pizza-margherita.jpg
        type: string
      price:
        example: "15000.00"
        type: string
      updated_at:
        example: "2022-01-01T00:00:00Z"
        type: string
    type: object
...
  model.UpdateGoodPriceRequest:
    properties:
      price:
        example: "15000.00"
        type: string
    required:
    - price
    type: object
  model.UpdateGoodRequest:
    properties:
      category_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      color_code:
        example: '#FF5733'
        type: string
      cook_time:
        example: 30
        type: integer
      department_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      description:
        example: Classic Italian pizza with tomatoes and mozzarella
        type: string
      description_i18n:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      name:
        example: Pizza Margherita
        type: string
      name_i18n:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      picture_url:
        example: https://example.com/pizza-margherita.jpg
        type: string
      price:
        example: "15000.00"
        type: string
    type: object
...
host: back.staging.maryai.yurtal.tech
info:
  contact: {}
  description: MaryAI API server with multi-language support (uz, ru, en)
  title: MaryAI API
  version: "1.0"
paths:
...
  /api/v1/categories/{category_id}/goods:
    get:
      description: Get goods by category with pagination
      parameters:
      - default: uz
        description: Language (uz, ru, en)
        in: query
        name: lang
        type: string
      - description: Category ID
        in: path
        name: category_id
        required: true
        type: string
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.GoodResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get goods by category
      tags:
      - Goods
...
  /api/v1/departments/{department_id}/goods:
    get:
      description: Get goods by department with pagination
      parameters:
      - default: uz
        description: Language (uz, ru, en)
        in: query
        name: lang
        type: string
      - description: Department ID
        in: path
        name: department_id
        required: true
        type: string
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.GoodResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get goods by department
      tags:
      - Goods
...
  /api/v1/goods:
    get:
      description: Get all goods with pagination
      parameters:
      - default: uz
        description: Language (uz, ru, en)
        in: query
        name: lang
        type: string
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.GoodResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all goods
      tags:
      - Goods
    post:
      consumes:
      - application/json
      description: Create a new good/menu item
      parameters:
      - default: uz
        description: Language (uz, ru, en)
        in: query
        name: lang
        type: string
      - description: Create good request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateGoodRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.GoodResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create good
      tags:
      - Goods
...
  /api/v1/goods/{id}:
    delete:
      description: Delete a good (soft delete)
      parameters:
      - default: uz
        description: Language (uz, ru, en)
        in: query
        name: lang
        type: string
      - description: Good ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete good
      tags:
      - Goods
    get:
      description: Get good by ID
      parameters:
      - default: uz
        description: Language (uz, ru, en)
        in: query
        name: lang
        type: string
      - description: Good ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GoodResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get good
      tags:
      - Goods
    put:
      consumes:
      - application/json
      description: Update good details
      parameters:
      - default: uz
        description: Language (uz, ru, en)
        in: query
        name: lang
        type: string
      - description: Good ID
        in: path
        name: id
        required: true
        type: string
      - description: Update good request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateGoodRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GoodResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update good
      tags:
      - Goods
  /api/v1/goods/{id}/price:
    put:
      consumes:
      - application/json
      description: Update good price
      parameters:
      - default: uz
        description: Language (uz, ru, en)
        in: query
        name: lang
        type: string
      - description: Good ID
        in: path
        name: id
        required: true
        type: string
      - description: Update good price request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateGoodPriceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GoodResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update good price
      tags:
      - Goods
  /api/v1/goods/{id}/restore:
    post:
      description: Restore a deleted good
      parameters:
      - default: uz
        description: Language (uz, ru, en)
        in: query
        name: lang
        type: string
      - description: Good ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GoodResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Restore good
      tags:
      - Goods
  /api/v1/goods/search:
    get:
      description: Search goods by name or description
      parameters:
      - default: uz
        description: Language (uz, ru, en)
        in: query
        name: lang
        type: string
      - description: Search query
        in: query
        name: query
        required: true
        type: string
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.GoodResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search goods
      tags:
      - Goods
  /api/v1/goods/search/by-price:
    get:
      description: Get goods within a price range with pagination
      parameters:
      - default: uz
        description: Language (uz, ru, en)
        in: query
        name: lang
        type: string
      - description: Minimum price
        in: query
        name: min_price
        required: true
        type: string
      - description: Maximum price
        in: query
        name: max_price
        required: true
        type: string
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.GoodResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get goods by price range
      tags:
      - Goods
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and your JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"